Rename WIH.
