+++ long-running-rest-workitem/assembly/assembly.xml
