+++ long-running-rest-workitem/src/main/java/org/jbpm/contrib/longrest/Constant.java
@@ -1,0 +1,17 @@
+package org.jbpm.contrib.longrest;
+
+/**
+ * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
+ */
+public class Constant {
+
+    public final static String CANCEL_SIGNAL_TYPE = "CancelAll";
+
+    public final static String OPERATION_FAILED_SIGNAL_TYPE = "operationFailed";
+
+    public static final String KIE_HOST_SYSTEM_PROPERTY = "KIE_HOSTNAME";
+
+    public static final String CANCEL_URL_JSON_POINTER_VARIABLE = "cancelUrlJsonPointer";
+
+    public static final String CANCEL_URL_TEMPLATE_VARIABLE = "cancelUrlTemplate";
+}
