+++ long-running-rest-workitem/src/main/java/org/jbpm/contrib/longrest/RemoteInvocationException.java
@@ -1,0 +1,12 @@
+package org.jbpm.contrib.longrest;
+
+public class RemoteInvocationException extends Exception {
+
+    public RemoteInvocationException(String message) {
+        super(message);
+    }
+
+    public RemoteInvocationException(String message, Exception e) {
+        super(message, e);
+    }
+}
