+++ long-running-rest-workitem/src/main/java/org/jbpm/contrib/longrest/ResponseProcessingException.java
@@ -1,0 +1,8 @@
+package org.jbpm.contrib.longrest;
+
+public class ResponseProcessingException extends Exception {
+
+    public ResponseProcessingException(String message, Exception e) {
+        super(message, e);
+    }
+}
