+++ long-running-rest-workitem/src/main/java/org/jbpm/contrib/longrest/WorkitemAbortedException.java
@@ -1,0 +1,5 @@
+package org.jbpm.contrib.longrest;
+
+public class WorkitemAbortedException extends Throwable {
+
+}
