+++ long-running-rest-workitem/src/main/java/org/jbpm/contrib/longrest/util/Maps.java
@@ -1,0 +1,11 @@
+package org.jbpm.contrib.longrest.util;
+
+import java.util.Map;
+
+public class Maps {
+
+    public static Map<String, Object> getStringObjectMap(Map<String, Object> map, String key) {
+        return (Map<String, Object>) map.get(key);
+    }
+
+}
