+++ long-running-rest-workitem/src/test/java/org/jbpm/contrib/RestServiceProcessEventListener.java
