+++ long-running-rest-workitem/src/test/java/org/jbpm/contrib/bpm/TestFunctions.java
