+++ long-running-rest-workitem/src/test/java/org/jbpm/contrib/demoservices/dto/BuildRequest.java
