--- rest-service-workitem/src/main/java/org/jbpm/contrib/restservice/Constant.java
@@ -1,17 +1,0 @@
-package org.jbpm.contrib.restservice;
-
-/**
- * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
- */
-public class Constant {
-
-    public final static String CANCEL_SIGNAL_TYPE = "CancelAll";
-
-    public final static String OPERATION_FAILED_SIGNAL_TYPE = "operationFailed";
-
-    public static final String KIE_HOST_SYSTEM_PROPERTY = "KIE_HOSTNAME";
-
-    public static final String CANCEL_URL_JSON_POINTER_VARIABLE = "cancelUrlJsonPointer";
-
-    public static final String CANCEL_URL_TEMPLATE_VARIABLE = "cancelUrlTemplate";
-}
