--- rest-service-workitem/src/main/java/org/jbpm/contrib/restservice/RemoteInvocationException.java
@@ -1,12 +1,0 @@
-package org.jbpm.contrib.restservice;
-
-public class RemoteInvocationException extends Exception {
-
-    public RemoteInvocationException(String message) {
-        super(message);
-    }
-
-    public RemoteInvocationException(String message, Exception e) {
-        super(message, e);
-    }
-}
