--- rest-service-workitem/src/main/java/org/jbpm/contrib/restservice/ResponseProcessingException.java
@@ -1,8 +1,0 @@
-package org.jbpm.contrib.restservice;
-
-public class ResponseProcessingException extends Exception {
-
-    public ResponseProcessingException(String message, Exception e) {
-        super(message, e);
-    }
-}
