--- rest-service-workitem/src/main/java/org/jbpm/contrib/restservice/WorkitemAbortedException.java
@@ -1,5 +1,0 @@
-package org.jbpm.contrib.restservice;
-
-public class WorkitemAbortedException extends Throwable {
-
-}
