--- rest-service-workitem/src/main/java/org/jbpm/contrib/restservice/util/Maps.java
@@ -1,11 +1,0 @@
-package org.jbpm.contrib.restservice.util;
-
-import java.util.Map;
-
-public class Maps {
-
-    public static Map<String, Object> getStringObjectMap(Map<String, Object> map, String key) {
-        return (Map<String, Object>) map.get(key);
-    }
-
-}
