--- rest-service-workitem/src/main/java/org/jbpm/contrib/restservice/util/ProcessUtils.java
@@ -1,66 +1,0 @@
-package org.jbpm.contrib.restservice.util;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.kie.api.runtime.KieSession;
-import org.kie.api.runtime.manager.RuntimeEngine;
-import org.kie.api.runtime.manager.RuntimeManager;
-import org.kie.api.runtime.process.WorkItem;
-import org.kie.api.runtime.process.WorkflowProcessInstance;
-import org.kie.internal.runtime.manager.context.ProcessInstanceIdContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
- */
-public class ProcessUtils {
-
-    public static WorkflowProcessInstance getProcessInstance(RuntimeManager runtimeManager, long processInstanceId) {
-        return (WorkflowProcessInstance)getKsession(runtimeManager, processInstanceId).getProcessInstance(processInstanceId);
-    }
-
-    public static KieSession getKsession(RuntimeManager runtimeManager, Long processInstanceId) {
-        if (runtimeManager != null) {
-            RuntimeEngine engine = runtimeManager.getRuntimeEngine(ProcessInstanceIdContext.get(processInstanceId));
-            return engine.getKieSession();
-        }
-        return null; //TODO
-    }
-
-    public static long getLongParameter(WorkItem workItem, String parameterName) {
-        Object parameter = workItem.getParameter(parameterName);
-        if (parameter != null) {
-            return Long.parseLong((String) parameter);
-        } else {
-            return -1L;
-        }
-    }
-
-    public static boolean getBooleanParameter(WorkItem workItem, String parameterName) {
-        Object parameter = workItem.getParameter(parameterName);
-        if (parameter != null) {
-            return (boolean) parameter;
-        } else {
-            return false;
-        }
-    }
-
-    public static String getStringParameter(WorkItem workItem, String parameterName) {
-        Object parameter = workItem.getParameter(parameterName);
-        if (parameter != null) {
-            return (String) parameter;
-        } else {
-            return "";
-        }
-    }
-
-    public static boolean getBooleanVariable(WorkflowProcessInstance processInstance, String name) {
-        Object variable = processInstance.getVariable(name);
-        if (variable instanceof Boolean) {
-            return (boolean) variable;
-        } else {
-            return Boolean.parseBoolean((String) variable);
-        }
-    }
-
-}
