--- rest-service-workitem/src/main/resources/RestServiceDefinitions.wid
@@ -1,35 +1,0 @@
-[
-        [
-            "name" : "RestServiceDefinitions",
-            "displayName" : "RestServiceDefinitions",
-            "category" : "rest-service-workitem",
-            "description" : "",
-            "defaultHandler" : "mvel: new org.jbpm.contrib.restservice.RestServiceWorkItemHandler(runtimeManager)",
-            "documentation" : "",
-
-            "parameters" : [
-                                "mustRunAfter" : new StringDataType()
-                ,"successCondition" : new StringDataType()
-                ,"taskTimeout" : new StringDataType()
-                ,"requestBody" : new StringDataType()
-                ,"requestUrl" : new StringDataType()
-                ,"requestMethod" : new StringDataType()
-                ,"cancelTimeout" : new StringDataType()
-                ,"alwaysRun" : new StringDataType()
-                ,"cancelUrlJsonPointer" : new StringDataType()
-
-            ],
-            "results" : [
-                                "result" : new StringDataType()
-
-            ],
-            "mavenDependencies" : [
-                                 "org.jbpm.contrib:rest-service-workitem:7.23.0.Final"
-                ,"org.slf4j:slf4j-api:"
-
-            ],
-            "icon" : ""
-
-        ]
-
-]
