--- rest-service-workitem/src/main/resources/RestTaskTimeout.wid
@@ -1,19 +1,0 @@
-[
-        [
-            "name" : "RestTaskTimeout",
-            "displayName" : "RestTaskTimeout",
-            "category" : "rest-service-workitem",
-            "description" : "",
-            "defaultHandler" : "mvel: new org.jbpm.contrib.restservice.TaskTimeoutWorkitemHandler(runtimeManager)",
-            "documentation" : "",
-
-            "mavenDependencies" : [
-                                 "org.jbpm.contrib:rest-service-workitem:7.23.0.Final"
-                ,"org.slf4j:slf4j-api:"
-
-            ],
-            "icon" : ""
-
-        ]
-
-]
